Fix types
Use double-precision everywhere.

+++ src/common.h
@@ -1,6 +1,6 @@
 #ifndef __COMMON_H__
 #define __COMMON_H__
 
-typedef float (*SynthCallback)(double time);
+typedef double (*SynthCallback)(double time);
 
 #endif
